backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-swagger-ui==4.11.1
python-dotenv==1.0.0
PyMuPDF==1.23.8
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10
sentence-transformers==2.2.2
chromadb==0.4.18
pydantic==2.5.2
requests==2.31.0